+++ src/openscm_runner/adapters/ciceroscm_py_adapter/read_results.py
@@ -5,12 +5,7 @@ and returns data to append to SCMRun
 import numpy as np
 import pandas as pd
 
-from ..utils._common_cicero_utils import (
-    fgas_list,
-    forc_sums,
-    ghg_not_fgas,
-    openscm_to_cscm_dict,
-)
+from ..utils._common_cicero_utils import get_data_from_forc_common, openscm_to_cscm_dict
 
 
 def get_data_from_conc(results, variable):
@@ -148,23 +143,13 @@ class CSCMREADER:
         Get data from forcing files
         """
         df_temp = results["forcing"]
-        years = df_temp.Year[:]
-        if variable in forc_sums:
-            timeseries = np.zeros(len(years))
-            for comp, value in self.variable_dict.items():
-                if variable in comp and value not in forc_sums:
-                    timeseries = timeseries + df_temp[value].to_numpy()
-        elif variable in ("Fgas", "GHG"):
-            timeseries = np.zeros(len(years))
-            for comp in fgas_list:
-                timeseries = timeseries + df_temp[comp].to_numpy()
-            if variable == "GHG":
-                for comp in ghg_not_fgas:
-                    timeseries = timeseries + df_temp[comp].to_numpy()
-        elif variable == "Total_forcing+sunvolc":
-            timeseries = df_temp["Total_forcing"].to_numpy()
-            timeseries = timeseries + self.get_volc_forcing(results)
-            timeseries = timeseries + self.get_sun_forcing(results)
-        else:
-            timeseries = df_temp[variable].to_numpy()
+        if variable == "Total_forcing+sunvolc":
+            volc = self.get_volc_forcing(results)
+            sun = self.get_sun_forcing(results)
+            return get_data_from_forc_common(
+                df_temp, variable, self.variable_dict, volc, sun
+            )
+        years, timeseries = get_data_from_forc_common(
+            df_temp, variable, self.variable_dict
+        )
         return years, timeseries
