+++ src/openscm_runner/adapters/utils/_common_cicero_utils.py
@@ -3,78 +3,7 @@ Module with various common utilities
 for both CICERO-SCM implementations
 to avoid code duplication
 """
-
-import openscm_units
-
-
-def _unit_conv_factor(unit, cicero_unit):
-    """
-    Convert cicero units that can't be automatically dealt with
-    in openscm_units
-    """
-    with openscm_units.unit_registry.context("NOx_conversions"):
-
-        if cicero_unit.startswith("GgH1"):
-            conv_factor = (
-                openscm_units.unit_registry(unit)
-                .to(cicero_unit.replace("GgH1", "GgHalon1"))
-                .magnitude
-            )
-        elif cicero_unit.startswith("GgH2"):
-            conv_factor = (
-                openscm_units.unit_registry(unit)
-                .to(cicero_unit.replace("GgH2", "GgHalon2"))
-                .magnitude
-            )
-        else:
-            conv_factor = openscm_units.unit_registry(unit).to(cicero_unit).magnitude
-
-    return conv_factor
-
-
-cicero_comp_dict = {
-    "CO2_lu": ["CO2|MAGICC AFOLU", 1],
-    "CFC-113": ["CFC113", 1],
-    "CFC-114": ["CFC114", 1],
-    "SO2": ["Sulfur", 1],
-    "NMVOC": ["VOC", 1],
-    "CFC-11": ["CFC11", 1],
-    "CFC-115": ["CFC115", 1],
-    "CFC-12": ["CFC12", 1],
-    "HCFC-141b": ["HCFC141b", 1],
-    "HCFC-142b": ["HCFC142b", 1],
-    "HCFC-22": ["HCFC22", 1],
-    "H-1211": ["Halon1211", 1],
-    "H-1301": ["Halon1301", 1],
-    "H-2402": ["Halon2402", 1],
-    "CO2": ["CO2|MAGICC Fossil and Industrial", 1],
-    "CH4": ["CH4", 1],
-    "N2O": ["N2O", 1],
-    "CH3Br": ["CH3Br", 1],
-    "CCl4": ["CCl4", 1],
-    "CH3CCl3": ["CH3CCl3", 1],
-    "HCFC-123": ["HCFC-123", 1],
-    "HFC125": ["HFC125", 1],
-    "HFC134a": ["HFC134a", 1],
-    "HFC143a": ["HFC143a", 1],
-    "HFC227ea": ["HFC227ea", 1],
-    "HFC23": ["HFC23", 1],
-    "HFC245fa": ["HFC245fa", 1],
-    "HFC32": ["HFC32", 1],
-    "HFC4310mee": ["HFC4310mee", 1],
-    "C2F6": ["C2F6", 1],
-    "C6F14": ["C6F14", 1],
-    "CF4": ["CF4", 1],
-    "SF6": ["SF6", 1],
-    "NOx": ["NOx", 1],
-    "CO": ["CO", 1],
-    "NH3": ["NH3", 1],
-    "BMB_AEROS_BC": ["BMB_AEROS_BC", 1],
-    "BMB_AEROS_OC": ["BMB_AEROS_OC", 1],
-    "BC": ["BC", 1],
-    "OC": ["OC", 1],
-}
-# Halon1212, CH3Cl
+import numpy as np
 
 openscm_to_cscm_dict = {
     "Surface Air Temperature Change": "dT_glob_air",
@@ -148,3 +77,32 @@ fgas_list = [
     "SF6",
 ]
 ghg_not_fgas = ["CO2", "CH4", "N2O", "TROP_O3", "STRAT_O3", "STRAT_H2O"]
+
+
+def get_data_from_forc_common(df_temp, variable, v_dict, volc=0, sun=0):
+    """
+    Get or calculate forcing when dataframe with forcers
+    variable and variable dictionary
+    If calculating with volcanic and solar forcing
+    methods to obtain these need to be supplied
+    """
+    years = df_temp.Year[:]
+    if variable in forc_sums:
+        timeseries = np.zeros(len(years))
+        for comp, value in v_dict.items():
+            if variable in comp and value not in forc_sums:
+                timeseries = timeseries + df_temp[value].to_numpy()
+    elif variable in ("Fgas", "GHG"):
+        timeseries = np.zeros(len(years))
+        for comp in fgas_list:
+            timeseries = timeseries + df_temp[comp].to_numpy()
+        if variable == "GHG":
+            for comp in ghg_not_fgas:
+                timeseries = timeseries + df_temp[comp].to_numpy()
+    elif variable == "Total_forcing+sunvolc":
+        timeseries = df_temp["Total_forcing"].to_numpy()
+        timeseries = timeseries + volc
+        timeseries = timeseries + sun
+    else:
+        timeseries = df_temp[variable].to_numpy()
+    return years, timeseries
