+++ tests/integration/test_ciceroSCM.py
@@ -8,10 +8,7 @@ from scmdata import ScmRun
 
 from openscm_runner import run
 from openscm_runner.adapters import CICEROSCM
-from openscm_runner.adapters.ciceroscm_adapter import (
-    make_scenario_files,
-    write_parameter_files,
-)
+from openscm_runner.adapters.ciceroscm_adapter import write_parameter_files
 from openscm_runner.utils import calculate_quantiles
 
 RTOL = 1e-5
@@ -301,32 +298,6 @@ class TestCICEROSCMAdapter(_AdapterTester):
                 output_variables=("Surface Air Temperature Change",),
                 out_config={"CiceroSCM": ("With ECS",)},
             )
-
-    def test_make_scenario_files(self, test_scenarios):
-        npt.assert_allclose(
-            3.0 / 11.0 * 1000.0,
-            make_scenario_files._unit_conv_factor("Gg CO2/yr", "Mg C/ yr"),
-        )
-        npt.assert_allclose(
-            28 / 44 / 1.0e12,
-            make_scenario_files._unit_conv_factor("kg N2O / yr", "Pg N2ON / yr"),
-        )
-        npt.assert_allclose(
-            14 / 46 / 1.0e12,
-            make_scenario_files._unit_conv_factor("kt NOx / yr", "Pt N / yr"),
-        )
-        npt.assert_allclose(
-            0.5,
-            make_scenario_files._unit_conv_factor("Tg SO2 / yr", "Tg S / yr"),
-        )
-        npt.assert_allclose(
-            1.0,
-            make_scenario_files._unit_conv_factor("Gg  Halon1211 / yr", "GgH1211 / yr"),
-        )
-        npt.assert_allclose(
-            1.0,
-            make_scenario_files._unit_conv_factor("Gg  Halon2402 / yr", "GgH2402 / yr"),
-        )
 
     @pytest.mark.parametrize(
         "input,exp",
