+++ tests/test-data/expected-integration-output/expected_ciceroscmpy_test_run_output.json
@@ -1,5 +1,5 @@
 {
-    "CICERO-SCM": [
+    "CICERO-SCM-PY": [
         [
             {
                 "variable": "Heat Content|Ocean",
